ai_agent/requirements.txt
google-adk
google-generativeai
sqlalchemy
psycopg2-binary
streamlit
python-multipart
google-cloud-aiplatform